chore(webui): fix typescript errors
