+++ webui/src/data/settings.ts
@@ -14,11 +14,8 @@ export async function getSettingsData() {
 	return { settingsData, defaultSettingsData, spotifyCredentials };
 }
 
-/**
- * @returns	{number}
- */
 export function getInitialPreviewVolume() {
-	let volume = parseInt(localStorage.getItem("previewVolume"));
+	let volume = parseInt(localStorage.getItem("previewVolume") ?? "");
 
 	if (isNaN(volume)) {
 		volume = 80; // Default
@@ -28,30 +25,18 @@ export function getInitialPreviewVolume() {
 	return volume;
 }
 
-/**
- * @returns	{boolean}
- */
 export function checkInitialSlimDownloads() {
 	return localStorage.getItem("slimDownloads") === "true";
 }
 
-/**
- * @returns	{boolean}
- */
 export function checkInitialSlimSidebar() {
 	return localStorage.getItem("slimSidebar") === "true";
 }
 
-/**
- * @returns	{boolean}
- */
 export function checkInitialShowBitrateTags() {
 	return localStorage.getItem("showBitrateTags") === "true";
 }
 
-/**
- * @returns	{boolean}
- */
 export function checkInitialShowSearchButton() {
 	return localStorage.getItem("showSearchButton") === "true";
 }
