+++ webui/tsconfig.json
@@ -11,7 +11,6 @@
 			"@/*": ["./src/*"]
 		}
 	},
-
 	"references": [
 		{
 			"path": "./tsconfig.config.json"
