+++ webui/vite.config.ts
@@ -1,7 +1,6 @@
 import { fileURLToPath, URL } from "node:url";
 
 import { defineConfig } from "vite";
-import svgLoader from "vite-svg-loader";
 import vue2 from "@vitejs/plugin-vue2";
 import vue2Jsx from "@vitejs/plugin-vue2-jsx";
